Create ofeakwu.html and update other recipes' file
The update on other recipes' file is on the 'title' tag.

The title tag update was crucial - which is simply the
addition of the word 'Recipes', followed by the actual
recipe name to be explained in that page.

With that addition, accessibility, ease of use, and
semantic rendering will be upheld with ease.